app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "horiyee", email = "horiyee.ka@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "dspy>=3.1.3",
    "python-dotenv>=1.2.1",
    "python-fasthtml>=0.12.42",
]

[project.scripts]
app = "app:main"
collect = "collector:main"
classify = "commit_prefix:main"

[build-system]
requires = ["uv_build>=0.9.28,<0.10.0"]
build-backend = "uv_build"
